+++ arch/loongarch/chunkset_lsx.c
@@ -40,7 +40,7 @@ static inline void storechunk(uint8_t *out, chunk_t *chunk) {
     __lsx_vst(*chunk, out, 0);
 }
 
-static inline chunk_t GET_CHUNK_MAG(uint8_t *buf, uint32_t *chunk_rem, uint32_t dist) {
+static inline chunk_t GET_CHUNK_MAG(uint8_t *buf, size_t *chunk_rem, size_t dist) {
     lut_rem_pair lut_rem = perm_idx_lut[dist - 3];
     __m128i perm_vec, ret_vec;
     /* Important to note:
